Add files via upload
This is the final project that we designed and developed for our freshman Computer Science course. It is a comprehensive student management system using C++ to improve the tracking and management of student data. We created a user-friendly student portal for data entry, retrieval, and analysis, as well as attendance tracking and grade management. We also developed an admin’s portal for managing student data. We also integrated user authentication and authorization features to ensure secure access to student data. Implemented a query system for students to register complaints and report issues to the administration team. The
Admin portal included a feature for administrators to receive and respond to student queries.